feat SmallPage: get rid of Basic version
